+++ examples/shared/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vultisig/examples-shared
 
+## 0.1.95
+
+### Patch Changes
+
+- Updated dependencies [[`f72cbc3`](https://github.com/vultisig/vultisig-sdk/commit/f72cbc35a23edb2b14984fce0a16495a3339e5e6), [`119d96d`](https://github.com/vultisig/vultisig-sdk/commit/119d96d5b2c9e1e2d8b322bf31d83f3ac4294244)]:
+  - @vultisig/sdk@2.15.0
+
 ## 0.1.94
 
 ### Patch Changes